requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dacite==1.6.0
dataclasses-json==0.5.7
Django==4.1.2
django-celery-results==2.4.0
django-environ==0.9.0
djangorestframework==3.14.0
kombu==5.2.4
marshmallow==3.18.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
pika==1.3.1
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-inspect==0.8.0
typing_extensions==4.4.0
vine==5.0.0
wcwidth==0.2.5
